pyproject.toml
[tool.poetry]
name = "jawafdehi"
version = "0.1.0"
description = ""
authors = ["Damodar Dahal <damo94761@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.8"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.9.0"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
# nepal-entity-service = {path = "../nes", extras = ["api", "scraping"], develop = true}
nepal-entity-service = {git = "https://github.com/Jawafdehi/NepalEntityService.git", extras = ["api", "scraping"]}
django-filter = "^25.2"
drf-spectacular = "^0.29.0"
dj-database-url = "^3.0.1"
django-jazzmin = "^3.0.1"
django-tinymce = "^5.0.0"
rules = "^3.5"
whitenoise = "^6.11.0"
django-auditlog = "^3.3.0"
gunicorn = "^23.0.0"
pydantic = "^2.12.5"
jsonpatch = "^1.33"
boto3 = "^1.35.0"
django-storages = "^1.14.3"
python-multipart = "^0.0.18"
djangorestframework-simplejwt = "^5.5.0"
anthropic = "^0.40.0"
langchain = {version = "^1.2.0", optional = true}
langchain-openai = {version = "^0.3.0", optional = true}
langchain-google-genai = {version = "^4.2.1", optional = true}
openai = {version = "^1.40.0", optional = true}

[tool.poetry.extras]
llm-all = ["langchain", "langchain-openai", "langchain-google-genai", "openai"]


[tool.poetry.group.dev.dependencies]
hypothesis = "^6.148.2"
pytest = "^9.0.1"
pytest-django = "^4.11.1"
pytest-asyncio = "^1.3.0"
black = "^26.3.0"
ruff = "^0.15.5"
pytest-timeout = "^2.4.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--reuse-db --timeout=10"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"